==> notebooks/01_acquire_player_game_logs.ipynb <==
# NBA 2024-25: Utilizing Roles
## Notebook 01: Data Acquisition
This notebook collects game-level player statistics for the 2024-25 NBA regular season using the NBA API. It combines all player game logs into a single DataFrame.

# Install NBA API
!pip install nba_api

# Import libraries
import pandas as pd
import time
from tqdm import tqdm
from nba_api.stats.endpoints import PlayerGameLog, leaguedashplayerstats
from nba_api.stats.static import players

___
## Acquiring Game-Level Player Statistics

# Get all players who registered a statistic during 2024-25 regular season
player_stats = leaguedashplayerstats.LeagueDashPlayerStats(
    season="2024-25",
    season_type_all_star="Regular Season"
).get_data_frames()[0]

print(f"{len(player_stats)} players")

# Player IDs of these 569 players
player_ids = player_stats["PLAYER_ID"].unique().tolist()

# --- Download each player's game log ---

all_game_logs = []

for pid in tqdm(player_ids, desc="Downloading player game logs"):
    try:
        game_logs = PlayerGameLog(player_id=str(pid), season="2024-25").get_data_frames()[0]
        game_logs["PLAYER_ID"] = pid
        all_game_logs.append(game_logs)
        time.sleep(0.6)
    except Exception as e:
        print(f"Failed: {pid}", e)

# Combine into one DataFrame
player_game_logs_2024_25 = pd.concat(all_game_logs, ignore_index=True)
player_game_logs_2024_25.head(5)

# Drop duplicate Player_ID column
player_game_logs_2024_25 = player_game_logs_2024_25.drop(columns=["PLAYER_ID"])

# Confirm columns
player_game_logs_2024_25.columns

> Player IDs are present, but player names are missing.

___
## Get Player Names

# Get player directory
players = pd.DataFrame(players.get_players())
players.head(5)

# Rename columns to match player logs column names
players = players.rename(columns={
    "id": "Player_ID",
    "full_name": "Player_Name"
})

# Merge player names with player game logs
player_game_logs_2024_25 = player_game_logs_2024_25.merge(
    players[["Player_ID", "Player_Name"]],
    on="Player_ID",
    how="left"
)

# Reorder columns so that `PLAYER_NAME` column is first
cols = ["Player_Name"] + [c for c in player_game_logs_2024_25.columns if c != "Player_Name"]
player_game_logs_2024_25 = player_game_logs_2024_25[cols]

# Inspect structure
player_game_logs_2024_25.head(5)

# Quick sanity checks
print(f"Number of rows: {len(player_game_logs_2024_25):,}")
print(f"Unique players: {player_game_logs_2024_25["Player_ID"].nunique()}")

# Drop nonessential columns
player_game_logs_2024_25 = player_game_logs_2024_25.drop(columns=["VIDEO_AVAILABLE"])

# Confirm updated columns
player_game_logs_2024_25.columns

___
## Save

# Save to Parquet
player_game_logs_2024_25.to_parquet("NBA_Player_GameLogs_2024_25.parquet")
