==> notebooks/02_acquire_team_game_logs.ipynb <==
# NBA 2024-25: Utilizing Roles
## Notebook 02: Data Acquisition, Pt. 2
This notebook collects game-level team statistics for the 2024-25 NBA regular season using the NBA API. It then merges them with game-level player statistics (from Notebook 01), so Usage Rate (USG%) and PRA (Points + Rebounds + Assists) can be calculated at the player-game level. 

# Install NBA API
!pip install nba_api

# Import libraries
import pandas as pd
from nba_api.stats.endpoints import leaguegamelog

# Ignore warnings
import warnings
warnings.filterwarnings("ignore")

___
## Load Player Game Logs

# Load game logs parquets (from Notebook 01)
player_logs = pd.read_parquet(r"C:\Users\dylan\OneDrive\Documents\Portfolio_Projects\NBA_2024_25_utilizing_roles\03_python_outputs\NBA_Player_GameLogs_2024_25.parquet")

___
## Acquire Team Game Logs

# --- Download each team's game logs ---
team_logs = leaguegamelog.LeagueGameLog(
    season="2024-25",
    season_type_all_star="Regular Season"
).get_data_frames()[0]

___
## Clean

# Player logs columns
player_logs.columns

# Team logs columns
team_logs.columns

# Only keep player columns necessary for USG% and Points-Rebounds-Assists calculations
player_logs_clean = player_logs[[
    "Player_Name", "Player_ID", "Game_ID", "GAME_DATE", "MATCHUP", "WL",
    "MIN", "FGM", "FGA", "FTA", "TOV", "REB", "AST", "PTS"
]].copy()

# Rename Game_ID column so matches the corresponding column in team logs
player_logs_clean = player_logs_clean.rename(columns={
    "Game_ID": "GAME_ID"
})

# Only keep team columns necessary for USG% calculation
team_logs_clean = team_logs[[
    "GAME_ID", "TEAM_ID", "TEAM_ABBREVIATION",
    "MIN", "FGA", "FTA", "TOV"
]].copy()

# Rename columns to emphasize they are team-related
team_logs_clean = team_logs_clean.rename(columns={
    "MIN": "TEAM_MIN",
    "FGA": "TEAM_FGA",
    "FTA": "TEAM_FTA",
    "TOV": "TEAM_TOV"
})

# Extract player's team abbreviation from MATCHUP
player_logs_clean["TEAM_ABBREVIATION"] = player_logs_clean["MATCHUP"].str.split().str[0]

# Confirm abbreviations align between team logs and player logs
set(player_logs_clean["TEAM_ABBREVIATION"]) - set(team_logs_clean["TEAM_ABBREVIATION"])

Empty set ✅

___
## Merge

# Merge player logs with team logs
merged = player_logs_clean.merge(
    team_logs_clean,
    on=["GAME_ID", "TEAM_ABBREVIATION"],
    how="left"
)

# Confirm columns
merged.columns

___
## Pre-Analysis Calculations

# Calculate USG% for every player-game combination
merged["USG%"] = (
    (merged["FGA"] + 0.44*merged["FTA"] + merged["TOV"]) * (merged["TEAM_MIN"] / 5)
) / (
    merged["MIN"] * (merged["TEAM_FGA"] + 0.44*merged["TEAM_FTA"] + merged["TEAM_TOV"])
)

merged["USG%"] = merged["USG%"] * 100

# Calculate PRA (Points + Rebounds + Assists) for every player-game combination
merged["PRA"] = merged["PTS"] + merged["REB"] + merged["AST"]

# Confirm
merged[["Player_Name", "GAME_DATE", "USG%", "PRA"]].sample(5)

___
## Save

# Save to parquet
merged.to_parquet("Merged_Player_Team_GameLogs_2024_25.parquet")
