==> notebooks/04_calculating_historical_all_star_baselines.ipynb <==
# NBA 2024-25: Utilizing Roles
## Notebook 04: Calculating All-Star Baselines (USG% and PRA)
This notebook flags players who were all-stars from the previous five seasons (2019-20 through 2023-24). It then calculates the **average Usage Rate (USG%)** and the **average PRA (Points + Rebounds + Assists)** for **every all-star during these five seasons**. These two averages are the "All-Star Baselines" that will become a major part of the later analysis.

# Import libraries
import pandas as pd
import numpy as np

___
## Load

# Load data
previous_5_seasons = pd.read_parquet(r"C:\Users\dylan\OneDrive\Documents\Portfolio_Projects\NBA_2024_25_utilizing_roles\03_python_outputs\AS_baselines\NBA_Per_Game_2019_2024.parquet")

# Calculate PRA (Points + Rebounds + Assists)
previous_5_seasons["PRA"] = previous_5_seasons["PTS"] + previous_5_seasons["TRB"] + previous_5_seasons["AST"]

# Confirm structure
previous_5_seasons.head(5)

___
## Flag All-Stars

# Flag all-stars
previous_5_seasons["is_allstar"] = previous_5_seasons["Awards"].str.contains("AS", na=False)

# Value counts
previous_5_seasons["is_allstar"].value_counts()

> There were **132 all-stars** from the previous 5 seasons.

___
## All-Star Baselines

# Filter for all-stars from previous 5 seasons
as_df = previous_5_seasons[previous_5_seasons["is_allstar"] == True].copy()

# Compute averages among these all-stars (and create a dictionary)
as_baseline = {
    "AS_USG_baseline": round(as_df["USG%"].mean(), 1),
    "AS_PRA_baseline": round(as_df["PRA"].mean(), 1)
}
as_baseline

> The average all-star USG% was 29.3% (per season).
> The average all-star PRA was 37.8 points + rebounds + assists (per game).

___
## Save

# Convert baseline dictionary to DataFrame
baseline_df = pd.DataFrame([
    {
        "Type": "All-Star",
        "USG_baseline": as_baseline["AS_USG_baseline"],
        "PRA_baseline": as_baseline["AS_PRA_baseline"]
    }
])
baseline_df

# Save to parquet
baseline_df.to_parquet("NBA_Per_Game_2019_2024_baselines.parquet")
